housekeeping: move metadata to widgets
