+++ src/view/frm_profile.py
@@ -12,7 +12,7 @@ from ..model.scratch_vector import ScratchVector
 from ..gp.feature_class_functions import import_existing, layer_path_parser
 from ..gp.import_temp_layer import ImportTemporaryLayer
 
-from .metadata import MetadataWidget
+from .widgets.metadata import MetadataWidget
 from .utilities import validate_name, add_standard_form_buttons
