+++ src/view/widgets/metadata.py
